requirements.txt
requests
azure-storage-blob
azure-storage-file-datalake
python-dotenv